Creature Template
Update creature templates to fit with a 1-60 level set up. Values gathered by comparing databases between AC and vMaNGOS.

If any creatures are removed from the world, they should be removed from this list too. They also need to be changed for a 60-70 set up and reset for a 70-80 set up.